requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
google-api-python-client==2.103.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
python-wordpress-xmlrpc==2.3
openai==0.28.1
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.0
celery==5.3.6
youtube-transcript-api==0.6.1